Move graph calculations to "calculator" classes
