+++ app/src/test/java/com/samco/trackandgraph/calculators/Statistics_calculateMovingAverages_KtTest.kt
@@ -15,7 +15,7 @@
  *  along with Track & Graph.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.samco.trackandgraph.statistics
+package com.samco.trackandgraph.calculators
 
 import com.samco.trackandgraph.database.entity.DataPoint
 import junit.framework.Assert.assertEquals
@@ -49,7 +49,7 @@ class Statistics_calculateMovingAverages_KtTest {
             val averagingDuration = Duration.ofHours(10)
 
             //WHEN
-            val answer = calculateMovingAverages(DataSample(dataPoints), averagingDuration)
+            val answer = MovingAverageCalculator(averagingDuration).execute(DataSample(dataPoints))
 
             //THEN
             val expected = listOf(5.0, 0.0, 2.0, 2.0, 1.5, 2.0, 8.0, 7.0, 3.0)
@@ -67,7 +67,7 @@ class Statistics_calculateMovingAverages_KtTest {
             val averagingDuration = Duration.ofHours(10)
 
             //WHEN
-            val answer = calculateMovingAverages(DataSample(dataPoints), averagingDuration)
+            val answer = MovingAverageCalculator(averagingDuration).execute(DataSample(dataPoints))
 
             //THEN
             assertEquals(0, answer.dataPoints.size)
